+++ plugin-dev/Source/Sentry/Private/Desktop/SentrySubsystemDesktop.h
@@ -37,6 +37,8 @@ private:
 	TSharedPtr<SentryCrashReporter> crashReporter;
 
 	bool isEnabled;
+
+	bool isStackTraceEnabled;
 };
 
 #endif