+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add stack trace attachment for regular events ([#301](https://github.com/getsentry/sentry-unreal/pull/301))
 - Add custom transport for Linux ([#297](https://github.com/getsentry/sentry-unreal/pull/297))
 
 ### Fixes